backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
beanie==1.26.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.1
pydantic[email]==2.7.4
httpx==0.27.0
aiosmtplib==3.0.1
python-dotenv==1.0.1
email-validator==2.1.1
Pillow==10.3.0
